CP-306_Базовая_проработка_JNI

+++ p7libapi/src/main/java/ru/petroplus/pos/p7LibApi/dto/DebitParamsDto.kt
@@ -13,8 +13,8 @@ package ru.petroplus.pos.p7LibApi.dto
 class DebitParamsDto(
     var serviceWhat: Int = 0,
     var serviceFrom: Int = 0,
-    var amount: Double = 0.0,
-    var price: Double = 0.0,
-    var sum: Double = 0.0,
+    var amount: UInt = 0u,
+    var price:  UInt = 0u,
+    var sum:    UInt = 0u,
     var pinBlock: ByteArray = byteArrayOf(),
 )
